requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Integrations
stripe==7.6.0
openai==1.3.7
tiktoken==0.5.2
httpx==0.25.2

# Caching & Queue
redis==5.0.1
celery==5.3.4

# CLI & Utils
click==8.1.7
rich==13.7.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0

# Code Quality
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Development
watchfiles==0.21.0
ipython==8.18.1